locations.py
https://en.wikipedia.org/wiki/Doppler_effect

Doppler factor calculation has an incorrect sign on line 138.  If the source and observer are moving closer together they will increase the frequency.  If the next range < current range,  source and observer moving closer together, then the range rate will be negative, which by would decrease the doppler factor not increase it. The sign change in line 138 is needed to correct this into the proper direction.